requirements.txt
Django
django-cms